Backend zack new

+++ database/migrations/2026_06_13_095534_create_events_table.php
@@ -11,27 +11,30 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('events', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('events')) {
+            Schema::create('events', function (Blueprint $table) {
+                $table->id();
 
-            $table->foreignId('admin_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('category_id')->constrained('event_categories')->onDelete('restrict');
+                //If an admin is deleted from the system, automatically delete all the events they created to prevent orphan data.
+                $table->foreignId('admin_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('category_id')->constrained('event_categories')->onDelete('restrict');
 
-            // Event details
-            $table->string('title', 255);
-            $table->text('description');
-            $table->string('venue', 255)->default('TBA');
-            $table->date('event_date');
-            $table->time('start_time');
-            $table->time('end_time');
-            $table->integer('maximum_slots');
-            $table->date('registration_deadline');
+                // Event details
+                $table->string('title', 255);
+                $table->text('description');
+                $table->string('venue', 255)->default('TBA');
+                $table->date('event_date');
+                $table->time('start_time');
+                $table->time('end_time');
+                $table->integer('maximum_slots');
+                $table->date('registration_deadline');
 
-            // enunm instead of string
-            $table->enum('status', ['Draft', 'Published', 'Cancelled', 'Completed'])->default('Draft');
+                // enunm instead of string
+                $table->enum('status', ['Draft', 'Published', 'Cancelled', 'Completed'])->default('Draft');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
